+++ src/core/resolve.js
@@ -1,5 +1,3 @@
-import 'es6-symbol/implement'
-import Map from 'es6-map'
 import assign from 'object-assign'
 
 import fs from 'fs'