+++ tests/src/package.js
@@ -1,5 +1,3 @@
-import 'es6-symbol/implement'
-
 var expect = require('chai').expect
 
 var path = require('path')