+++ src/core/getExports.js
@@ -1,6 +1,3 @@
-import 'es6-symbol/implement'
-import Map from 'es6-map'
-
 import * as fs from 'fs'
 
 import { createHash } from 'crypto'