+++ src/rules/no-namespace.js
@@ -1,0 +1,17 @@
+/**
+ * @fileoverview Rule to disallow namespace import
+ * @author Radek Benkel
+ */
+
+//------------------------------------------------------------------------------
+// Rule Definition
+//------------------------------------------------------------------------------
+
+
+module.exports = function (context) {
+  return {
+    'ImportNamespaceSpecifier': function (node) {
+      context.report(node, `Unexpected namespace import.`)
+    },
+  }
+}