+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 ## [Unreleased]
 ### Added
 - report resolver errors at the top of the linted file
+- add `no-namespace` rule
 
 ## [1.4.0] - 2016-03-25
 ### Added